+++ test/e2e/helper_test.go
@@ -1326,3 +1326,99 @@ func OrgNUser(n int) (username, userID string, err error) {
 
 	return users.GetResults()[n].Username, users.GetResults()[n].GetId(), nil
 }
+
+func deleteKeys(t *testing.T, cliPath string, toDelete map[string]struct{}) {
+	t.Helper()
+
+	cmd := exec.Command(cliPath,
+		orgEntity,
+		"apiKeys",
+		"ls",
+		"-o=json")
+
+	cmd.Env = os.Environ()
+	resp, err := RunAndGetStdOut(cmd)
+	require.NoError(t, err, string(resp))
+
+	var keys atlasv2.PaginatedApiApiUser
+	err = json.Unmarshal(resp, &keys)
+	require.NoError(t, err)
+
+	uniqueKeysToDelete := make(map[string]struct{})
+	for _, key := range keys.GetResults() {
+		keyID := key.GetId()
+		desc := key.GetDesc()
+
+		if _, ok := toDelete[desc]; ok {
+			uniqueKeysToDelete[keyID] = struct{}{}
+		}
+	}
+
+	for keyID := range uniqueKeysToDelete {
+		errs := []error{}
+		t.Logf("Deleting key with ID: %s", keyID)
+		cmd = exec.Command(cliPath,
+			orgEntity,
+			"apiKeys",
+			"rm",
+			keyID,
+			"--force")
+		cmd.Env = os.Environ()
+		_, err = RunAndGetStdOutAndErr(cmd)
+		if err != nil && !strings.Contains(err.Error(), "API_KEY_NOT_FOUND") {
+			errs = append(errs, err)
+		}
+		if len(errs) > 0 {
+			t.Errorf("unexpected errors while deleting keys: %v", errs)
+		}
+	}
+}
+
+func deleteOrgInvitations(t *testing.T, cliPath string) {
+	t.Helper()
+	cmd := exec.Command(cliPath,
+		orgEntity,
+		invitationsEntity,
+		"ls",
+		"-o=json")
+	cmd.Env = os.Environ()
+	resp, err := RunAndGetStdOut(cmd)
+	t.Logf("%s\n", resp)
+	require.NoError(t, err, string(resp))
+	var invitations []atlasv2.OrganizationInvitation
+	require.NoError(t, json.Unmarshal(resp, &invitations), string(resp))
+	for _, i := range invitations {
+		deleteOrgInvitation(t, cliPath, *i.Id)
+	}
+}
+
+func deleteOrgTeams(t *testing.T, cliPath string) {
+	t.Helper()
+
+	cmd := exec.Command(cliPath,
+		teamsEntity,
+		"ls",
+		"-o=json")
+	cmd.Env = os.Environ()
+	resp, err := RunAndGetStdOut(cmd)
+	t.Logf("%s\n", resp)
+	require.NoError(t, err, string(resp))
+	var teams atlasv2.PaginatedTeam
+	require.NoError(t, json.Unmarshal(resp, &teams), string(resp))
+	for _, team := range teams.GetResults() {
+		assert.NoError(t, deleteTeam(team.GetId()))
+	}
+}
+
+func deleteOrgInvitation(t *testing.T, cliPath string, id string) {
+	t.Helper()
+	cmd := exec.Command(cliPath,
+		orgEntity,
+		invitationsEntity,
+		"delete",
+		id,
+		"--force")
+	cmd.Env = os.Environ()
+	resp, err := RunAndGetStdOut(cmd)
+	require.NoError(t, err, string(resp))
+}
